requirements.txt
Flask==3.0.0
flask-socketio==5.3.5
python-dotenv==1.0.0
mysql-connector-python==8.2.0
python-engineio==4.8.0
python-socketio==5.10.0
Werkzeug==3.0.1
eventlet==0.33.3